Lucius_Scoring/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.9.2
PyMuPDF==1.24.9
google-generativeai==0.8.3
tqdm==4.66.5
google-genai==0.3.0
